pyproject.toml
[tool.poetry]
name = "tiko"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.2"
drf-spectacular = "^0.27.2"
django-environ = "^0.11.2"


[tool.poetry.group.dev.dependencies]
pytest-django = "^4.8.0"
factory-boy = "^3.3.0"
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tiko.settings"
